+++ backend/budget/models.py
@@ -1,9 +1,8 @@
 from django.db import models
+from django.utils.translation import gettext_lazy as _
 
 from common.models import IndexedTimeStampedModel
 from users.models import User
-
-from django.utils.translation import gettext_lazy as _
 
 
 class Transaction(IndexedTimeStampedModel):