+++ backend/budget/views/expense.py
@@ -1,13 +1,13 @@
 from __future__ import annotations
 
+from budget.handlers.add_new_transaction import AddNewTransactionHandler
+from budget.requests.add_new_transaction import AddNewTransactionRequest
+from budget.serializer import ExpenseSerializer
 from rest_framework import viewsets
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
-from budget.handlers.add_transaction import AddTransactionRequestManager
-from budget.requests.add_transaction import AddTransactionRequest
-from budget.serializer import ExpenseSerializer
 
 
 class ExpenseAddViewSet(viewsets.ViewSet):
@@ -17,12 +17,12 @@ class ExpenseAddViewSet(viewsets.ViewSet):
         detail=False,
         methods=['post'],
         permission_classes=[IsAuthenticated],
-        url_path='expense',  # TODO: Maybe change the path to budget/<int:budget_id>/expense
+        url_path='budget/expense',  # TODO: Maybe change the path to budget/<int:budget_id>/expense
     )
     def add(self, request: Request) -> Response:
-        expense_request_manager = AddTransactionRequestManager(
+        expense_request_manager = AddNewTransactionHandler(
             rest_request=request,
-            factory=AddTransactionRequest,
+            factory=AddNewTransactionRequest,
             serializer=ExpenseSerializer,
             transaction_type='expenses',
         )