+++ backend/users/views/users.py
@@ -1,20 +1,19 @@
 from __future__ import annotations
 
-from auth.handlers.check_auth import CheckAuthRequestManager
-from auth.requests.email_token import EmailTokenRequest
 from rest_framework import viewsets
 from rest_framework.decorators import action
 from rest_framework.permissions import AllowAny
 from rest_framework.request import Request
+from users.handlers.get_all import GetAllUsersRequestManager
 
 
-class CheckAuthViewSet(viewsets.ViewSet):
+class UsersViewSet(viewsets.ViewSet):
     @action(
         detail=False,
-        methods=['post'],
+        methods=['get'],
         permission_classes=[AllowAny],
-        url_path='auth/check',
+        url_path='users/all',
     )
-    def check_auth(self, request: Request):
-        check_auth_request_manager = CheckAuthRequestManager(rest_request=request)
+    def get_all(self, request: Request):
+        check_auth_request_manager = GetAllUsersRequestManager(rest_request=request)
         return check_auth_request_manager.safe_process().response()